Use same table resize strategy for both concurrent maps
I don't see a benefit for storing an additional field on the
TableEntry for ConcurrentLong2ReferenceChainedHashTable and this
brings both maps closer to the same implementation.

Additionally, change all bin node retrievals to use acquire ordering
instead of volatile. The table is retrieved using volatile ordering,
which is all we need to guarantee the total ordering of the map.