Ensure queued task in PrioritisedTaskQueue is always present in queue
Previously when the priority was adjusted we would remove the
task from the queue and then re-insert the task. This left a window
where the task did not appear to exist inside of the queue.

Now we insert the new task holder when adjusting the priority
before removing the old holder, ensuring that the task is
always present in queue while it is queued.

This fixes the worker threads from BalancedPrioritisedThreadPool
potentially parking forever while a single task exists in queue,
as the worker threads assume that once there are no queues with
any tasks available to peek that the queues are empty.

See https://github.com/PaperMC/Paper/issues/13275